Handle case where modules share memory in ModuleEventAllocMonitor
Previous tested showed that some groups of modules use their own
memory caching which led to deallocations at end of event being
assigned to the wrong module (as they happen to have used that
address during the event as well but was later deallocated by
a different module). Looking for module matches in reverse order
of module completion helps with proper assignment.